split arima and auto arima
